View count api added, type edits

+++ lib/model/top_favorite_model.dart
@@ -32,7 +32,7 @@ class Result {
   dynamic id;
   String companyName;
   String companyLogo;
-  int? commentsAvg;
+  double? commentsAvg;
   String? companyAddress;
   int? viewsNumber;
